细节优化

+++ albedo-boot-ui-angular/src/main/webapp/app/theme/pages/sys/role/role.component.ts
@@ -3,6 +3,7 @@ import {ScriptLoaderService} from "../../../../shared/base/service/script-loader
 import {CTX, DATA_STATUS, DICT_SYS_DATA} from "../../../../app.constants";
 import {SessionStorageService} from "ngx-webstorage";
 import {ActivatedRoute} from "@angular/router";
+import {Principal} from "../../../../auth/_services/principal.service";
 
 declare let datatable: any;
 @Component({
@@ -15,8 +16,8 @@ export class RoleComponent implements OnInit,OnDestroy, AfterViewInit {
 
     routerSub: any;
     constructor(private _script: ScriptLoaderService,
-        private router: ActivatedRoute,
-        private sessionStorage: SessionStorageService) {
+                private principal: Principal,
+        private router: ActivatedRoute) {
 
     }
 
@@ -40,6 +41,7 @@ export class RoleComponent implements OnInit,OnDestroy, AfterViewInit {
     }
 
     initTable() {
+        var thisPrincipal = this.principal;
         // const token = this.localStorage.retrieve('authenticationToken') || this.sessionStorage.retrieve('authenticationToken');
         var options = {
             data: {
@@ -90,18 +92,23 @@ export class RoleComponent implements OnInit,OnDestroy, AfterViewInit {
                     sortable: false,
                     overflow: 'visible',
                     template: function(row) {
-                        return '\
-						<a href="#/sys/role/form/'+ row.id + '" class="m-portlet__nav-link btn m-btn m-btn--hover-accent m-btn--icon m-btn--icon-only m-btn--pill" title="编辑">\
-							<i class="la la-edit"></i>\
-						</a>\
-						<a href="javascript:void(0)" class="m-portlet__nav-link btn m-btn m-btn--hover-warning m-btn--icon m-btn--icon-only m-btn--pill confirm" title="'+ (row.status == "正常" ? "锁定" : "解锁") + '角色"\
-						 data-table-id="#data-table-role" data-method="put"  data-title="你确认要操作【'+ row.name+ '】角色吗？" data-url="'+ CTX +'/sys/role/'+ row.id+ '">\
-							<i class="la la-'+ (row.status == "正常" ? "unlock-alt" : "unlock") + '"></i>\
-						</a>\
-					    <a href="javascript:void(0)" class="m-portlet__nav-link btn m-btn m-btn--hover-danger m-btn--icon m-btn--icon-only m-btn--pill confirm" title="删除"\
-                             data-table-id="#data-table-role" data-method="delete"  data-title="你确认要删除【'+ row.name+ '】角色吗？" data-url="'+ CTX +'/sys/role/'+ row.id+ '">\
-                            <i class="la la-trash"></i>\
-                        </a>';
+
+                        var template = '';
+                        if (thisPrincipal.hasAuthority("sys_role_edit"))
+                            template += '<a href="#/sys/role/form/' + row.id + '" class="m-portlet__nav-link btn m-btn m-btn--hover-accent m-btn--icon m-btn--icon-only m-btn--pill" title="编辑">\
+                                \<i class="la la-edit"></i>\
+                                \</a>';
+                        if (thisPrincipal.hasAuthority("sys_role_lock"))
+                            template += '<a href="javascript:void(0)" class="m-portlet__nav-link btn m-btn m-btn--hover-warning m-btn--icon m-btn--icon-only m-btn--pill confirm" title="' + (row.status == "正常" ? "锁定" : "解锁") + '角色"\
+						 data-table-id="#data-table-role" data-method="put"  data-title="你确认要操作【' + row.name + '】角色吗？" data-url="' + CTX + '/sys/role/' + row.id + '">\
+                                \<i class="la la-\'+ (row.status == "正常" ? "unlock-alt" : "unlock") + \'"></i>\
+                                \</a>';
+                        if (thisPrincipal.hasAuthority("sys_role_delete"))
+                            template += '<a  href="javascript:void(0)" class="m-portlet__nav-link btn m-btn m-btn--hover-danger m-btn--icon m-btn--icon-only m-btn--pill confirm" title="删除"\
+                                   data-table-id="#data-table-role" data-method="delete"  data-title="你确认要删除【' + row.name + '】角色吗？" data-url="'+ CTX +'/sys/role/'+ row.id+ '">\
+                                \<i class="la la-trash"></i>\
+                                \</a>';
+                        return template;
                     },
                 }],
         };
